Implementa servidor unificado para o sistema de controle de tarefas, integrando React e Node.js. Atualiza scripts de inicialização, remove diretórios de backend e frontend, e ajusta a estrutura do projeto para simplificar o deployment e a manutenção. Melhora a documentação no README.md e atualiza as dependências no package.json.
