--- backend/SENHA POSTGREE.txt
@@ -1,1 +1,0 @@
-86f3331f3c02482082c1c0e5e9816b02
