--- backend/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "analise-dados-backend",
-  "version": "1.0.0",
-  "description": "Backend para o sistema de controle de tarefas.",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "pg": "^8.7.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.15"
-  }
-}
