--- frontend/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "analise-dados-frontend",
-  "version": "0.1.0",
-  "private": true,
-  "proxy": "http://localhost:3001",
-  "dependencies": {
-    "@ag-grid-community/client-side-row-model": "^31.3.1",
-    "@ag-grid-community/core": "^31.3.1",
-    "@ag-grid-community/csv-export": "^31.3.1",
-    "@ag-grid-community/react": "^31.3.1",
-    "ag-grid-community": "^31.3.4",
-    "ag-grid-react": "^31.3.4",
-    "axios": "^0.21.4",
-    "bootstrap": "^5.3.6",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.7.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^5.5.0",
-    "react-scripts": "5.0.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.5.1",
-    "vite": "^6.3.5"
-  }
-}
