+++ package.json
@@ -1,17 +1,57 @@
 {
-  "name": "analise-dados-react",
+  "name": "sistema-controle-tarefas",
   "version": "1.0.0",
-  "description": "Sistema de Controle de Tarefas com React e Node.js",
-  "main": "index.js",
+  "description": "Sistema completo de controle de tarefas com React e Node.js em um servidor único",
+  "main": "server/server.js",
+  "homepage": ".",
   "scripts": {
-    "install-backend": "cd backend && npm install",
-    "install-frontend": "cd frontend && npm install",
-    "install-all": "npm run install-backend && npm run install-frontend",
-    "start-backend": "cd backend && npm start",
-    "start-frontend": "cd frontend && npm start",
-    "dev": "concurrently \"npm run start-backend\" \"npm run start-frontend\""
+    "start": "node server/server.js",
+    "dev": "concurrently \"npm run build:watch\" \"npm run server:dev\"",
+    "server:dev": "nodemon server/server.js",
+    "build": "react-scripts build",
+    "build:watch": "react-scripts build --watch",
+    "test": "react-scripts test",
+    "install-deps": "npm install",
+    "setup": "npm run install-deps && npm run build"
   },
   "dependencies": {
-    "concurrently": "^7.6.0"
+    "@ag-grid-community/client-side-row-model": "^31.3.1",
+    "@ag-grid-community/core": "^31.3.1",
+    "@ag-grid-community/csv-export": "^31.3.1",
+    "@ag-grid-community/react": "^31.3.1",
+    "ag-grid-community": "^31.3.4",
+    "ag-grid-react": "^31.3.4",
+    "axios": "^0.21.4",
+    "bootstrap": "^5.3.6",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "pg": "^8.7.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.5.0",
+    "react-scripts": "5.0.1"
+  },
+  "devDependencies": {
+    "concurrently": "^7.6.0",
+    "nodemon": "^2.0.15"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
