+++ server/SENHA POSTGREE.txt
@@ -1,0 +1,1 @@
+86f3331f3c02482082c1c0e5e9816b02
